@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -17,13 +17,17 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @author     Sven Baumann <baumann.sv@gmail.com>
+ * @copyright  2012-2022 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
 namespace MetaModels\AttributeTableMultiBundle\Test\Attribute;
 
+use Contao\CoreBundle\Framework\Adapter;
+use Contao\StringUtil;
+use Contao\Validator;
 use Doctrine\DBAL\Connection;
 use MetaModels\Attribute\IAttributeTypeFactory;
 use MetaModels\AttributeTableMultiBundle\Attribute\AttributeTypeFactory;
@@ -32,6 +36,8 @@ use PHPUnit\Framework\TestCase;
 
 /**
  * Test the attribute factory.
+ *
+ * @covers \MetaModels\AttributeTableMultiBundle\Attribute\AttributeTypeFactory
  */
 class TableMultiAttributeTypeFactoryTest extends TestCase
 {
@@ -87,9 +93,11 @@ class TableMultiAttributeTypeFactoryTest extends TestCase
      */
     protected function getAttributeFactories()
     {
-        $connection  = $this->mockConnection();
+        $connection = $this->mockConnection();
+        $stringUtil = new Adapter(StringUtil::class);
+        $validator  = new Adapter(Validator::class);
 
-        return array(new AttributeTypeFactory($connection));
+        return array(new AttributeTypeFactory($connection, $stringUtil, $validator));
     }
 
     /**
@@ -99,9 +107,11 @@ class TableMultiAttributeTypeFactoryTest extends TestCase
      */
     public function testCreateAttribute()
     {
-        $connection  = $this->mockConnection();
+        $connection = $this->mockConnection();
+        $stringUtil = new Adapter(StringUtil::class);
+        $validator  = new Adapter(Validator::class);
 
-        $factory   = new AttributeTypeFactory($connection);
+        $factory   = new AttributeTypeFactory($connection, $stringUtil, $validator);
         $values    = array(
             'tabletext_cols' => ''
         );
