@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,19 +18,25 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @author     Sven Baumann <baumann.sv@gmail.com>
+ * @copyright  2012-2022 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
 namespace MetaModels\AttributeTableMultiBundle\Test\Attribute;
 
+use Contao\CoreBundle\Framework\Adapter;
+use Contao\StringUtil;
+use Contao\Validator;
 use Doctrine\DBAL\Connection;
 use MetaModels\AttributeTableMultiBundle\Attribute\TableMulti;
 use PHPUnit\Framework\TestCase;
 
 /**
  * Unit tests to test class TableMulti.
+ *
+ * @covers \MetaModels\AttributeTableMultiBundle\Attribute\TableMulti
  */
 class TableMultiTest extends TestCase
 {
@@ -84,7 +90,16 @@ class TableMultiTest extends TestCase
      */
     public function testInstantiation()
     {
-        $text = new TableMulti($this->mockMetaModel('en', 'en'), [], $this->mockConnection());
+        $stringUtil = new Adapter(StringUtil::class);
+        $validator  = new Adapter(Validator::class);
+
+        $text = new TableMulti(
+            $this->mockMetaModel('en', 'en'),
+            [],
+            $this->mockConnection(),
+            $stringUtil,
+            $validator
+        );
         $this->assertInstanceOf('MetaModels\Attribute\TableMulti\TableMulti', $text);
     }
 }
