@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_tablemulti.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,7 +15,8 @@
  * @author     Andreas Dziemba <adziemba@web.de>
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2022 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_tablemulti/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -29,7 +30,7 @@ use PHPUnit\Framework\TestCase;
 /**
  * This class tests if the deprecated autoloader works.
  *
- * @package MetaModels\AttributeTableMultiBundle\Test
+ * @covers \MetaModels\AttributeTableMultiBundle\DeprecatedAutoloader
  */
 class DeprecatedAutoloaderTest extends TestCase
 {
